requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.0.3
numpy==1.24.3
pydantic==2.4.2
httpx==0.25.0
google-cloud-bigquery==3.20.1
google-auth==2.29.0
db-dtypes==1.2.0
